requirements.txt
openai
streamlit_chat
streamlit-folium
geopy